requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
aksharamukha==2.3
alembic==1.13.1
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
Deprecated==1.2.18
distro==1.9.0
filelock==3.18.0
Flask==2.3.3
flask-cors==6.0.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
fonttools==4.58.2
frozenlist==1.6.0
fsspec==2025.5.0
google-auth==2.40.1
google-genai==1.15.0
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.31.4
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
jaconv==0.4.0
Jinja2==3.1.3
jiter==0.10.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
langcodes==3.5.0
language_data==1.3.0
litellm==1.70.2
lxml==5.4.0
Mako==1.3.10
marisa-trie==1.2.1
MarkupSafe==3.0.2
multidict==6.4.4
openai==1.75.0
packaging==25.0
propcache==0.3.1
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.4
pydantic_core==2.33.2
pykakasi==2.3.0
python-dotenv==1.1.0
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rpds-py==0.25.1
rsa==4.9.1
setuptools==80.9.0
sniffio==1.3.1
SQLAlchemy==2.0.25
tiktoken==0.9.0
tokenizers==0.21.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
unicodedata2==16.0.0
urllib3==2.4.0
websockets==15.0.1
Werkzeug==2.3.7
wrapt==1.17.2
yarl==1.20.0
zipp==3.21.0
